requirements.txt
# anthropic>=0.18.0
# langchain>=0.1.0
# langgraph>=0.0.20
# python-dotenv>=1.0.0
# tweepy>=4.14.0
# fastapi>=0.109.0
# uvicorn>=0.27.0
# pydantic>=2.5.0
# httpx>=0.26.0


fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.0
python-dotenv==1.0.0
tweepy==4.14.0